packages/pipeline/pyproject.toml
[project]
name = "maparchive-pipeline"
version = "0.1.0"
description = "STAC catalog generation pipeline for the CIESIN Map Archive"
requires-python = ">=3.12"
dependencies = [
    "pystac>=1.12.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
maparchive = "maparchive_pipeline.cli:main"

[dependency-groups]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
